Move Frame._concat to DataFrame._concat since it's a DataFrame-specific implementation.
